requirements-analysis.txt
# Requirements for video analysis query script
psycopg2-binary>=2.9.0